wip project tidy up + adding startup static library
