+++ Game/random.h
@@ -1,7 +1,7 @@
 #ifndef HEADER_HH_RANDOM
 #define HEADER_HH_RANDOM
 
-#include "types.h"
+#include "game_types.h"
 
 // Random numbers between 0 and 1,000,000,000 in base 16
 // @see https://www.random.org/integers/?mode=advanced
