+++ Game/utility.h
@@ -1,7 +1,7 @@
 #ifndef HEADER_HH_UTILITIES
 #define HEADER_HH_UTILITIES
 
-#include "types.h"
+#include "game_types.h"
 #include "intrinsics.h"
 
 uint32 u32RoundUpDivide(uint32 dividend, uint32 divisor);
