+++ Game/world.h
@@ -1,7 +1,7 @@
 #ifndef HEADER_HH_WORLD
 #define HEADER_HH_WORLD
 
-#include "types.h"
+#include "game_types.h"
 #include "tilemap.h"
 
 // How many pixels to represent 1 meter?
