+++ Handmade Hero.sln
@@ -6,6 +6,7 @@ MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Platform Windows", "Platform Windows\Platform Win32.vcxproj", "{271FCAD7-A0D8-4D27-85C6-510A29F543DB}"
 	ProjectSection(ProjectDependencies) = postProject
 		{39BEC1AA-0EE5-4B0B-AC79-E9D8D000B61F} = {39BEC1AA-0EE5-4B0B-AC79-E9D8D000B61F}
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4} = {997DFED0-4464-406C-900F-FFFA1ADA20A4}
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{02516E6B-0E49-4519-BE52-E3A8FA0F719D}"
@@ -18,6 +19,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Game", "Game\Game.vcxproj", "{39BEC1AA-0EE5-4B0B-AC79-E9D8D000B61F}"
+	ProjectSection(ProjectDependencies) = postProject
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4} = {997DFED0-4464-406C-900F-FFFA1ADA20A4}
+	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Startup", "Startup\Startup.vcxproj", "{997DFED0-4464-406C-900F-FFFA1ADA20A4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,6 +49,14 @@ Global
 		{39BEC1AA-0EE5-4B0B-AC79-E9D8D000B61F}.Release|x64.Build.0 = Release|x64
 		{39BEC1AA-0EE5-4B0B-AC79-E9D8D000B61F}.Release|x86.ActiveCfg = Release|Win32
 		{39BEC1AA-0EE5-4B0B-AC79-E9D8D000B61F}.Release|x86.Build.0 = Release|Win32
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4}.Debug|x64.ActiveCfg = Debug|x64
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4}.Debug|x64.Build.0 = Debug|x64
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4}.Debug|x86.ActiveCfg = Debug|Win32
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4}.Debug|x86.Build.0 = Debug|Win32
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4}.Release|x64.ActiveCfg = Release|x64
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4}.Release|x64.Build.0 = Release|x64
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4}.Release|x86.ActiveCfg = Release|Win32
+		{997DFED0-4464-406C-900F-FFFA1ADA20A4}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
