+++ Startup/x64/Debug/Startup.log
@@ -1,0 +1,2 @@
+﻿  startup.cpp
+  Startup.vcxproj -> C:\Users\jonca\Documents\clones\handmde\x64\Debug\Startup.lib
