+++ Startup/x64/Debug/Startup.tlog/Cl.items.tlog
@@ -1,0 +1,1 @@
+C:\Users\jonca\Documents\clones\handmde\Startup\startup.cpp;C:\Users\jonca\Documents\clones\handmde\Startup\x64\Debug\startup.obj
