+++ Startup/x64/Debug/Startup.tlog/Startup.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.36.32532:TargetPlatformVersion=10.0.22000.0:
+Debug|x64|C:\Users\jonca\Documents\clones\handmde\|
